requirements.txt
amqplib==1.0.2
lockfile==0.9.1
python-daemon==1.6
setuptools
